refactor: move action-menu key handler into handlers/keys (#494)
Next slice of the handlers/ extraction. handle_action_menu_key moves from an
inline impl App method into handlers::keys as a free function over &mut App,
with a one-line delegation shim in app.rs. The dispatcher itself is small; the
bulk (execute_action_by_hint) stays in app.rs and is promoted to pub(crate) so
the moved handler can call it, alongside the already-pub action_menu_items.

ListKeyAction is no longer pattern-matched in app.rs, so it is dropped from the
list_overlay import there. Behavior unchanged.

No App fields changed; field-count ratchet unaffected.
